explore tristate
There seems to be a regime at very low gamma where the system oscillate between 2 states: one of regular rivalry and another were activity is relatively balanced. So a total of 3 states. Here I make some more phase-space plots  at this regime, but the results are not striking at all.